Separated Match and Country to a separate class for reuse and added initial Bets UI
