+++ lib/bet.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/cupertino.dart';
+import 'match.dart';
 
 class Bet extends StatelessWidget {
   Bet({@required this.phase});
@@ -7,6 +8,35 @@ class Bet extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Text("Bet");
+    return Container(
+      margin: EdgeInsets.all(14),
+      child: Column(
+          mainAxisSize: MainAxisSize.min,
+          children: <Widget>[
+            Text("Apuestas",style: TextStyle(fontWeight: FontWeight.w700, fontSize: 20)),
+            Container(
+                height: MediaQuery.of(context).size.height*0.7,
+                child: Match(
+              host: 'Bolivia',
+              guest: 'Argentina',
+              hostFlag:
+                  'http://ssl.gstatic.com/onebox/media/sports/logos/SGxeD7yhwPj53FmPBmMMHg_48x48.png',
+              guestFlag:
+                  'http://ssl.gstatic.com/onebox/media/sports/logos/1xBWyjjkA6vEWopPK3lIPA_48x48.png',
+              date: '2020-10-13',
+              hour: '15:00',
+              result: '1:2',
+              showOtherBets: false
+            ),
+              ),
+            CupertinoButton.filled(
+              child: Text('Apostar'),
+              onPressed: () { 
+                /** Save Bets Action */
+               },
+            ),
+          ],
+        )
+    );
   }
 }
